# api/openapi/wallet.yaml
openapi: 3.1.0
info:
  title: Exchange wallet API
  version: 0.2.0
  description: |
    Deposits and withdrawals (requirements §5.10, §11.5, §11.6). Each user gets one deposit
    address per network. Most networks are served by the custody wallet
    (ADR-0011: USDT on TRC20, BEP20 and ERC20, BTC, ETH): the custodian
    creates the address, watches the chain and reports a deposit once it
    is confirmed; on the platform's own network (ETH on Sepolia) the
    address is derived from the platform's deposit key and the scanner
    follows the chain. Either way a deposit is credited to the spot
    account once it has the network's confirmations. Deposits below the
    network's minimum, or to a closed account, are not credited to the
    user but held for manual handling (UNCLAIMED_DEPOSIT); transfers of
    tokens the platform does not list are recorded as REJECTED and not
    booked. Changes are pushed on the private WebSocket channel
    "deposits".

    Withdrawals go to addresses of the caller's address book, usable after
    a cooling-off period (24 hours; shorter in the test environment).
    Adding an address and requesting a withdrawal each take a step-up
    (X-Step-Up-Token from POST /v1/auth/step-up; with an authenticator app
    bound, proven by it). A request is checked (eligibility, minimum,
    daily and monthly limits in USDT), scored by the risk rules and its
    amount plus fee frozen; it is then APPROVED, or PENDING_REVIEW when a
...
    (ADR-0008).
servers:
...
security:
  - bearer: []
...
paths:
  /v1/wallet/networks:
    get:
      operationId: listWalletNetworks
      summary: The networks an asset moves on
      description: |
        What the deposit and withdrawal pages show of each network: its
        name, address format, confirmations, usual time to a credit,
        minimums, withdrawal fee and explorer links. USDT has one balance
        whatever network it arrives on; the network is chosen per deposit
        address and per withdrawal.
      tags: [deposits]
      parameters:
        - name: asset
          in: query
          description: The asset (case-insensitive); every asset's networks when omitted.
          schema:
            type: string
            examples: [USDT]
      responses:
        '200':
          description: The networks, closed ones included (see the switches).
          content:
            application/json:
              schema:
                type: object
                required: [networks]
                properties:
                  networks:
                    type: array
                    items:
                      $ref: '#/components/schemas/WalletNetwork'
        default:
          $ref: './common.yaml#/components/responses/Error'
  /v1/wallet/limits:
    get:
      operationId: getWithdrawLimits
      summary: The caller's withdrawal limits in effect
      description: |
        The daily and monthly limits in USDT the next withdrawal is checked
        against, what is used of them (today and this month, UTC), and the
        full limits: both identities (email and phone) and an authenticator
        app give them, the app only once it has been bound for
        totp_settling_hours (24); until then full_limits_at says when.
        The pages read the numbers and the hours from here.
      tags: [withdrawals]
      responses:
        '200':
          description: The limits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawLimits'
        default:
          $ref: './common.yaml#/components/responses/Error'
  /v1/wallet/withdraw-addresses/validate:
    post:
      operationId: validateWithdrawAddress
      summary: Check an address before saving or using it
      description: |
        Checks the form for the network (EVM: 0x and 40 hex digits, EIP-55
        when mixed case; TRON: Base58Check starting with T; Bitcoin:
        bech32/bech32m or Base58Check of the right network) and whether it
        is a platform deposit address: the caller's own is refused, another
        user's makes the withdrawal an internal transfer without a network
        fee. The form is all it can tell: a valid address may still belong
        to nobody.
      tags: [withdrawals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [network, address]
              properties:
                asset:
                  type: string
                  description: Picks the network's asset when several share it.
                network:
                  type: string
                  examples: [ETH-SEPOLIA]
                address:
                  type: string
                memo:
                  type: string
                  description: Memo or tag, for networks that need one.
      responses:
        '200':
          description: The outcome; an invalid address is not an error.
          content:
            application/json:
              schema:
                type: object
                required: [valid, network, address_format, normalized, reason, internal]
                properties:
                  valid:
                    type: boolean
                  network:
                    type: string
                  address_format:
                    type: string
                    enum: [EVM, TRON, BTC]
                  normalized:
                    type: [string, 'null']
                    description: The address as the network writes it (EIP-55 case for EVM); null when invalid.
                  reason:
                    type: [string, 'null']
                    enum: [ADDRESS_FORMAT, ADDRESS_CHECKSUM, ADDRESS_NETWORK, MEMO_REQUIRED, ADDRESS_OWN, ADDRESS_RETIRED, null]
                    description: |
                      Why it is invalid: not an address of the network's kind,
                      a checksum mismatch (a typo), an address of another
                      network (e.g. testnet), a missing memo, the caller's
                      own deposit address, or a deposit address taken out of
                      use (the custodian's stand-in made it up: no chain
                      knows it).
                  internal:
                    type: boolean
                    description: Another user's deposit address; the withdrawal completes inside the platform with no fee.
        default:
          $ref: './common.yaml#/components/responses/Error'
  /v1/wallet/deposit-address:
    get:
      operationId: getDepositAddress
      summary: The caller's deposit address for an asset on a network
      description: |
        Assigns the next derived address on first use; the same address
        serves every asset of the network. Fails with
        WALLET_NETWORK_UNKNOWN for a network the asset does not have,
        WALLET_NETWORK_DISABLED while the asset or the network takes no
        deposits, WALLET_UNAVAILABLE before the deposit key is configured,
        and the USER_ eligibility codes.
      tags: [deposits]
      parameters:
        - name: asset
          in: query
          required: true
          schema:
            type: string
            examples: [ETH]
        - name: network
          in: query
          required: true
          schema:
            type: string
            examples: [ETH-SEPOLIA]
      responses:
        '200':
          description: The address.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositAddress'
        default:
          $ref: './common.yaml#/components/responses/Error'
  /v1/wallet/deposits:
    get:
      operationId: listDeposits
      summary: The caller's deposits, newest first
      tags: [deposits]
      parameters:
        - name: cursor
          in: query
          description: next_cursor of the previous page.
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: One page.
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deposit'
                  next_cursor:
                    type: [string, 'null']
        default:
          $ref: './common.yaml#/components/responses/Error'
  /v1/wallet/withdraw-addresses:
    get:
      operationId: listWithdrawAddresses
      summary: The caller's withdrawal address book
      tags: [withdrawals]
      responses:
        '200':
          description: The entries, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WithdrawAddress'
        default:
          $ref: './common.yaml#/components/responses/Error'
    post:
      operationId: addWithdrawAddress
      summary: Add an address to the book
      description: |
        Needs a step-up. The address can be used from usable_at on. An
        address already in the book is returned as it is. Fails with
        WALLET_INVALID_ADDRESS, WALLET_NETWORK_UNKNOWN, WALLET_OWN_ADDRESS
        (the caller's own deposit address) or AUTH_STEP_UP_REQUIRED.
      tags: [withdrawals]
      parameters:
        - $ref: '#/components/parameters/StepUp'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [network, address]
              additionalProperties: false
              properties:
                network:
                  type: string
                  examples: [ETH-SEPOLIA]
                address:
                  type: string
                label:
                  type: string
                  maxLength: 50
      responses:
        '201':
          description: The entry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawAddress'
        default:
          $ref: './common.yaml#/components/responses/Error'
  /v1/wallet/withdraw-addresses/{id}:
    delete:
      operationId: deleteWithdrawAddress
      summary: Remove an address from the book
      tags: [withdrawals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Removed.
        default:
          $ref: './common.yaml#/components/responses/Error'
...
  /v1/wallet/withdrawals/{id}:
    get:
      operationId: getWithdrawal
      summary: One of the caller's withdrawals
      tags: [withdrawals]
      parameters:
        - $ref: '#/components/parameters/WithdrawalID'
      responses:
        '200':
          description: The withdrawal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Withdrawal'
        default:
          $ref: './common.yaml#/components/responses/Error'
    delete:
      operationId: cancelWithdrawal
      summary: Cancel a withdrawal that is not signed yet
      description: |
        Releases the frozen funds. Fails with WALLET_WITHDRAWAL_NOT_CANCELABLE
        once it is being signed or is with the custodian.
      tags: [withdrawals]
      parameters:
        - $ref: '#/components/parameters/WithdrawalID'
      responses:
        '200':
          description: The withdrawal, CANCELED.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Withdrawal'
        default:
          $ref: './common.yaml#/components/responses/Error'
...
components:
  parameters:
    StepUp:
      name: X-Step-Up-Token
      in: header
      required: true
      description: A step-up token from POST /v1/auth/step-up; each works once.
      schema:
        type: string
    WithdrawalID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WalletNetwork:
      type: object
      required: [asset, network, display_name, chain, address_format, contract, confirmations, eta_minutes, min_deposit, min_withdraw,
        withdraw_fee, memo_required, deposit_enabled, withdraw_enabled, withdraw_suspended, explorer_tx_url, explorer_address_url]
      properties:
        asset:
          type: string
        network:
          type: string
          examples: [ETH-SEPOLIA]
        display_name:
          type: string
          examples: [TRC20, ERC20, Sepolia]
        chain:
          type: string
        address_format:
          type: string
          enum: [EVM, TRON, BTC]
        contract:
          type: [string, 'null']
          description: The token contract; null for the chain's native coin.
        confirmations:
          type: integer
        eta_minutes:
          type: integer
          description: Usual minutes from the transfer to the credit; 0 when unknown.
        min_deposit:
          type: string
        min_withdraw:
          type: string
        withdraw_fee:
          type: string
        memo_required:
          type: boolean
        deposit_enabled:
          type: boolean
        withdraw_enabled:
          type: boolean
          description: The network takes withdrawals now; false while the asset's withdrawals are suspended.
        withdraw_suspended:
          type: boolean
          description: >-
            The asset's withdrawals are suspended while the platform checks its funds (requests fail with
            WALLET_WITHDRAW_SUSPENDED); deposits go on.
        explorer_tx_url:
          type: [string, 'null']
          description: Explorer link of a transaction with a {tx} placeholder.
        explorer_address_url:
          type: [string, 'null']
          description: Explorer link of an address with an {address} placeholder.
...